# docs/swagger.yaml
basePath: /
info:
  contact: {}
  description: router.org clone implemented with Go + Gin by Elif ERDEN.
  title: HTTPBin Clone
  version: "1.0"
paths:
  /absolute-redirect/{n}:
    get:
      description: Performs absolute redirects
      parameters:
      - description: Redirect count
        in: path
        name: "n"
        required: true
        type: integer
      responses:
        "302":
          description: Found
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Absolute redirect
      tags:
      - response
...
  /bytes/{n}:
    get:
      description: Returns N random bytes
      parameters:
      - description: Number of bytes
        in: path
        name: "n"
        required: true
        type: integer
      produces:
      - application/octet-stream
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Generate random bytes
      tags:
      - response
  /cache/{n}:
    get:
      description: Returns response with Cache-Control header
      parameters:
      - description: Cache duration in seconds
        in: path
        name: "n"
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Cache response
      tags:
      - response
  /cookies:
    get:
      description: Returns all cookies sent by the client
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              additionalProperties:
                type: string
              type: object
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get cookies
      tags:
      - cookies
  /cookies/set/{name}/{value}:
    get:
      description: Sets a cookie and redirects to /cookies
      parameters:
      - description: Cookie name
        in: path
        name: name
        required: true
        type: string
      - description: Cookie value
        in: path
        name: value
        required: true
        type: string
      responses:
        "302":
          description: Found
          headers:
            Location:
              description: /cookies
              type: string
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Set a cookie
      tags:
      - cookies
...
  /delay/{n}:
    get:
      description: Delays response by N seconds
      parameters:
      - description: Delay in seconds
        in: path
        name: "n"
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: integer
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delay response
      tags:
      - response
...
  /hidden-basic-auth/{user}/{passwd}:
    get:
      description: Same as basic auth but endpoint is hidden
      parameters:
      - description: Username
        in: path
        name: user
        required: true
        type: string
      - description: Password
        in: path
        name: passwd
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Hidden Basic Authentication
      tags:
      - auth
...
  /image/jpeg:
    get:
      description: Returns a sample JPEG image
      produces:
      - image/jpeg
      responses:
        "200":
          description: OK
          schema:
            type: file
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get JPEG image
      tags:
      - images
  /image/webp:
    get:
      description: Returns a sample WEBP image
      produces:
      - image/webp
      responses:
        "200":
          description: OK
          schema:
            type: file
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get WEBP image
      tags:
      - images
...
  /links/{n}:
    get:
      description: Generates N HTML links
      parameters:
      - description: Number of links
        in: path
        name: "n"
        required: true
        type: integer
      produces:
      - text/html
      responses:
        "200":
          description: HTML links
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Generate links
      tags:
      - misc
...
  /redirect/{n}:
    get:
      description: Performs N redirects ending at /get
      parameters:
      - description: Redirect count
        in: path
        name: "n"
        required: true
        type: integer
      responses:
        "302":
          description: Found
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Redirect N times
      tags:
      - response
  /relative-redirect/{n}:
    get:
      description: Performs relative redirects
      parameters:
      - description: Redirect count
        in: path
        name: "n"
        required: true
        type: integer
      responses:
        "302":
          description: Found
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Relative redirect
      tags:
      - response
...
  /status/{code}:
    get:
      description: Returns the given status code with empty body (except teapot).
      parameters:
      - description: HTTP status code
        in: path
        maximum: 599
        minimum: 100
        name: code
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            type: string
        "204":
          description: No Content
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "418":
          description: I'm a teapot
          schema:
            type: string
      summary: Return a status code
      tags:
      - response
...
schemes:
- http
swagger: "2.0"